airflow/requirements.txt
gdown
pyarrow
pymongo
python-dotenv
pandas
scipy
scikit-learn
pip-tools
mlflow
dvc
dvc_gdrive
pre-commit
apache-airflow
altair
evidently
numpy
pillow
python-dateutil
streamlit
types-python-dateutil
typing_extensions
optuna
